RF_07.000.00 PeopleIconSizeTest deleted, tests moved to PeopleTest. Deleted FreestyleProject21Test
